requirements.txt
Flask==3.1.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.8.0
python-dotenv==1.0.1
bcrypt==4.3.0
python-jose==3.5.0
